requirements.txt
google-generativeai==0.3.0
pdfplumber==0.10.3
chromadb==0.4.24
sentence-transformers==2.2.2
numpy==1.24.3
click==8.1.7
python-dotenv==1.0.0
faiss-cpu==1.7.4
